% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{recode_system_size}
\alias{recode_system_size}
\title{Recode system time variable}
\usage{
recode_system_size(data)
}
\arguments{
\item{data}{dataframe containing data from file
`gillespie.csv`}
}
\value{
`data` with variable `system_size` recoded for
plotting
}
\description{
Recode system time variable in dataframe containing data from file
`gillespie.csv`
}
\examples{
{
data <- readr::read_csv(system.file("testdata", "gillespie.csv",
package = "rrcompendium"))
recode_system_size(data)
}
}
